feat :seperate reducer for functionallity
