+++ src/reducer/index.js
@@ -1,1 +1,2 @@
 export { reducer } from "./reducer";
+export { filterReducer } from "./filterReducer";
